quantum-fraud-detection/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
imbalanced-learn>=0.12.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0.1
xgboost>=2.0.0

# Qiskit packages
qiskit>=1.4.4
qiskit-aer>=0.17.2
qiskit-machine-learning>=0.8.4
qiskit-algorithms>=0.4.0
qiskit-ibm-runtime>=0.30.0
